requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
DateTime==4.3
Deprecated==1.2.13
finnhub-python==2.4.7
idna==3.3
nexmo==2.5.2
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
urllib3==1.26.8
wrapt==1.13.3
zope.interface==5.4.0
